seperated the feed into a new component instead of having it all in homepage
